GNZ-817 Auto-delete `/tmp` folders for dart projects and `genezio local`
Previously, there were a lot of temporary folders created if
dart projects were deployed using `genezio deploy` or if
any kind of project was tested with `genezio local`.

Fixed this by removing all those folders in case of errors.

Also, updated the name of the temporary folders to start
with `genezio-` prefix, instead of `foo-`.

Signed-off-by: Costin-Robert Sin <sin.costinrobert@gmail.com>