Enable `noPropertyAccessFromIndexSignature` TypeScript compiler option (#598)
<!--
  Before submitting a Pull Request, please ensure you've done the following:
  - Read the Genezio Contributing Guide: https://github.com/Genez-io/genezio/blob/master/CONTRIBUTING.md
  - Read the Genezio Code of Conduct: https://github.com/Genez-io/genezio/blob/master/CODE_OF_CONDUCT.md
  - Provide tests for your changes.
  - Add comments on your code.
  - Use descriptive commit messages.
  - Update any related documentation and include any relevant screenshots for UI/UX changes.
-->

## Type of change

<!-- Please delete options that are not relevant. -->

-   [x] 🧑‍💻 Improvement

## Description

This change disallows developers to use `dot access` on properties that are declared as IndexSignature;

Example:

```ts
interface Foo {
  a: number;
  b: string;
  [key: string]: string;
}

function func(foo: Foo) {
  foo.a;             // allowed
  foo.b;             // allowed
  foo.anything;      // not allowed
  foo["anything"];   // allowed
}
```

This TS Compiler rule helps other ESLint rules not trigger in the future, when we enable a stricter typed ESLint rule set.

<!--
Please do not leave this blank. Add a small summary of the PR:

This PR [adds/removes/fixes/replaces] the [feature/bug/etc].

List any dependencies that are required for this change.
-->

## Checklist

-   [x] My code follows the contributor guidelines of this project;
-   [x] New and existing unit tests pass locally with my changes;