Reimplement the entryfile detection algorithm
Now, we are searching the package.json for a main module.
If this is found in the filesystem, we are setting that as an entryfile.
If the user defined and used the main file as set in package.json, we
will trust he set it correctly.

Else, we are searching for patterns and set it as an entryfile.

This approach is less error-prone.

Signed-off-by: Andreia Ocănoaia <andreia.ocanoaia@gmail.com>